pyproject.toml
[project]
name = "aia"
version = "0.1.0"
description = "Accuracy Intelligence Agent scaffold"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.8.0",
  "langgraph>=0.2.0",
  "fastapi>=0.112.0",
  "uvicorn>=0.30.0",
  "python-multipart>=0.0.9",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
